requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-logging==3.11.3
Werkzeug==3.0.3
google-cloud-dns==0.34.0
google-auth==2.36.0
python-dotenv==0.20.0
